+++ src/components/exercises/ExercisesCategories/ExercisesCategories.jsx
@@ -2,20 +2,12 @@ import React, { useEffect } from 'react';
 import { NavLink } from 'react-router-dom';
 import styles from './ExercisesCategories.module.css';
 import ExercisesList from '../ExercisesList';
-import { useSelector, useDispatch } from 'react-redux';
-import { selectCategories } from 'redux/exercises/selectors';
-import { exerciseCategories } from 'redux/exercises/operation';
 import { useExercise } from 'hooks';
 
+
+
 function ExercisesCategories() {
-  const { categories } = useExercise();
-  const dispatch = useDispatch();
-
-  console.log(categories)
-
-  useEffect(() => {
-    dispatch(exerciseCategories('bodyparts'));
-  }, [dispatch]);
+  const { selectedCategory } = useExercise();
 
   return (
     <>
@@ -26,7 +18,7 @@ function ExercisesCategories() {
         <h2 className={styles.exercisesSubTitle}>Exercises</h2>
         <ul className={styles.exercisesList}>
           <li className={styles.exercisesItem}>
-            <NavLink to="/" className={styles.exercisesLinkActive}>
+            <NavLink to="bodyparts" className={styles.exercisesLinkActive}>
               Body parts
             </NavLink>
           </li>
