+++ src/components/exercises/ExercisesItem/ExercisesItem.module.css
@@ -1,61 +1,46 @@
 .exercisesContainer {
+  /* max-width: 237px;
+  width: 100%; */
+  aspect-ratio: 237 / 206;
+  border-radius: 12px;
+  background-position: center;
+  background-size: cover;
+  /* @media screen and (min-width: 1440px) {
+    height: 206px;
+  } */
+}
+
+.text_container {
   display: flex;
   flex-direction: column;
   width: 100%;
-  height: 141px;
-  padding: 16px;
+  height: 100%;
+  background: linear-gradient(
+    0deg,
+    rgba(4, 4, 4, 0.5) 0%,
+    rgba(4, 4, 4, 0.5) 100%
+  );
+  color: white;
   justify-content: center;
-  align-items: stretch;
+  align-items: center;
   border-radius: 12px;
   border: 1px solid var(--transparent-medium-3);
-  background: var(--transparent-medium-5);
-  margin-top: 20px;
 }
 
-@media screen and (min-width: 768px) {
-  .exercisesCard {
-    width: calc((100% - 16px) / 3);
-    overflow: hidden;
-  }
+.text_name {
+  color: #efede8;
+  font-size: 24px;
+  font-weight: 400;
+  line-height: 1.3;
 }
 
-@media screen and (min-width: 1440px) {
-  .exercisesCard {
-    max-width: 405px;
-  }
+.text_name::first-letter {
+  text-transform: capitalize;
 }
 
-.exercisesCard {
-  box-sizing: border-box;
+.text {
+  color: rgba(239, 237, 232, 0.4);
+  font-size: 12px;
+  font-weight: 400;
+  line-height: 1.5;
 }
-
-/* .excercisesCard {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  height: 141px;
-  padding: 16px;
-  justify-content: center;
-  align-items: stretch;
-  border-radius: 12px;
-  border: 1px solid var(--transparent-medium-3);
-  background: var(--transparent-medium-5);
-  margin-top: 20px;
-}
-
-@media screen and (min-width: 768px) {
-  .exercisesCard {
-    width: calc((100% - 16px) / 2);
-    overflow: hidden;
-  }
-}
-
-@media screen and (min-width: 1440px) {
-  .exercisesCard {
-    max-width: 405px;
-  }
-}
-
-.exercisesCard {
-  box-sizing: border-box;
-} */
