+++ src/components/exercises/ExercisesList/ExercisesList.jsx
@@ -1,18 +1,27 @@
 import { useEffect, useState } from 'react';
-import { useDispatch } from 'react-redux';
-import { useExercises } from '../../../hooks/exercises';
 import ExercisesItem from '../ExercisesItem';
 import styles from './ExercisesList.module.css';
-import { exercisesList } from '../../../redux/exercises/operation';
-import { useAuth } from '../../../hooks/auth';
-import BasicModalWindow from '../../BasicModalWindow';
+import { useDispatch } from 'react-redux';
+import { exerciseCategories } from 'redux/exercises/operation';
+import { useExercise } from 'hooks';
+
 
 
 function ExercisesList() {
+  const { categories } = useExercise();
+  const dispatch = useDispatch();
+  const currentCategory = 'equipment';
+
+  console.log(categories)
+
+  useEffect(() => {
+    dispatch(exerciseCategories('bodyparts'));
+  }, [dispatch]);
+
   return (
     <ul className={styles.exercisesList}>
-      <ExercisesItem />
-    </ul>
+      {categories.map((category) => <ExercisesItem item={category.title}/>)}
+   </ul>
   );
 }
